notification etc changes

+++ neo4j-services/src/main/java/com/campusconnect/neo4j/da/GroupDao.java
@@ -1,16 +1,21 @@
 package com.campusconnect.neo4j.da;
 
+import com.campusconnect.neo4j.da.iface.NotificationDao;
 import com.campusconnect.neo4j.da.iface.UserDao;
 import com.campusconnect.neo4j.da.mapper.DBMapper;
 import com.campusconnect.neo4j.da.utils.FilterHelper;
+import com.campusconnect.neo4j.da.utils.TargetHelper;
 import com.campusconnect.neo4j.repositories.GroupRepository;
 import com.campusconnect.neo4j.repositories.UserGroupRepository;
+import com.campusconnect.neo4j.types.common.Target;
 import com.campusconnect.neo4j.types.neo4j.Book;
 import com.campusconnect.neo4j.types.neo4j.Group;
 import com.campusconnect.neo4j.types.neo4j.User;
 import com.campusconnect.neo4j.types.neo4j.UserGroupRelationship;
 import com.campusconnect.neo4j.types.web.GroupMember;
+import com.campusconnect.neo4j.types.web.Notification;
 import com.campusconnect.neo4j.util.Constants;
+
 import org.neo4j.rest.graphdb.entity.RestNode;
 import org.neo4j.rest.graphdb.entity.RestRelationship;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,6 +42,9 @@ public class GroupDao {
     private UserDao userDao;
 
     private Neo4jTemplate neo4jTemplate;
+    
+    @Autowired
+    NotificationDao notificationDao;
 
     public GroupDao(Neo4jTemplate neo4jTemplate) {
         this.neo4jTemplate = neo4jTemplate;
@@ -48,10 +56,33 @@ public class GroupDao {
     public Group createGroup(Group group) {
         return groupRepository.save(group);
     }
+    
+    public void deleteGroup(String groupId)
+    {
+    	  Group group = getGroup(groupId);
+    	  groupRepository.delete(group);
+    	  
+    }
 
-    public void deleteGroup(String groupId) {
+    public void deleteGroupByAdmin(String groupId,String userId) {
         Group group = getGroup(groupId);
-        groupRepository.delete(group);
+        
+        UserGroupRelationship existingUserGroupRelationship = userGroupRepository.getUserGroupRelationShip(userId, groupId);
+        if(existingUserGroupRelationship.getRole().equalsIgnoreCase(Constants.GROUP_ADMIN_ROLE))
+        {
+        		List<GroupMember> groupMembers = getMembers(groupId, userId);
+        		String groupName = group.getName();
+        		
+        		groupRepository.delete(group);
+        		
+        		for(GroupMember groupMember:groupMembers)
+        		{
+        			String groupMemberId = groupMember.getId();
+        			Target target = TargetHelper.createDeleteGroupTarget(groupName);
+        			Notification notification =  new Notification(target, System.currentTimeMillis(), Constants.GROUP_DELETED_NOTIFICATION_TYPE);
+        			notificationDao.addNotification(groupMemberId, notification);
+        		}
+        }
     }
 
     public Group getGroup(String groupId) {
@@ -101,18 +132,33 @@ public class GroupDao {
         if (null != existingUserGroupRelationship) {
             existingUserGroupRelationship.setRole(role);
             neo4jTemplate.save(existingUserGroupRelationship);
+            //TODO : Add check if role is admin if yes then check created by is admin
+            //TODO : else why is this request coming ??
+            
         } else {
             Long currentTime = System.currentTimeMillis();
             UserGroupRelationship userGroupRelationship = new UserGroupRelationship(
                     createdBy, currentTime, group, currentTime,
                     role, user);
+            
+            Target target = TargetHelper.createGroupTarget(group);
+            Notification notification = new Notification(target, System.currentTimeMillis(), Constants.GROUP_MEMBER_ADDED);
+            //TODO :Add in notification type of user role like admin or member 
+            
+            notificationDao.addNotification(userId, notification);
+            
             neo4jTemplate.save(userGroupRelationship);
         }
+        
+        
     }
 
-    public Group updateGroup(String groupId, Group group) {
+    public Group updateGroup(String groupId, Group group, String userId) {
         Group groupToBeUpdated = getGroup(groupId);
         groupToBeUpdated.setName(group.getName());
+        groupToBeUpdated.setLastModifiedBy(userId);
+        groupToBeUpdated.setLastModifiedTime(System.currentTimeMillis());
+        //TODO : Should notification besent that group name has been changed to ALL USER ??
 
         return groupRepository.save(groupToBeUpdated);
     }
@@ -124,4 +170,59 @@ public class GroupDao {
     public List<Book> getWishListBooks(String groupId) {
         return groupRepository.getWishListBooks(groupId);
     }
+
+	public void exitFromGroup(String groupId, String userId) {
+		
+		 Group group = getGroup(groupId);
+	        User user = userDao.getUser(userId);
+	        UserGroupRelationship existingUserGroupRelationship = userGroupRepository.getUserGroupRelationShip(userId, groupId);
+	        
+	        if(null != existingUserGroupRelationship)
+	        {
+	        	if(existingUserGroupRelationship.getRole().equalsIgnoreCase(Constants.GROUP_MEMBER_ROLE))
+	        	{
+	        		userGroupRepository.delete(existingUserGroupRelationship);
+	        	}
+	        	else if(existingUserGroupRelationship.getRole().equalsIgnoreCase(Constants.GROUP_ADMIN_ROLE))
+	        	{
+	        		List<GroupMember> groupMembers = getAllGroupUsers(groupId);
+	        		boolean adminExists =  false;
+	        		
+	        		for(GroupMember groupMember:groupMembers)
+	        		{
+	        			if(groupMember.getRole().equalsIgnoreCase(Constants.GROUP_ADMIN_ROLE))
+	        			{
+	        				adminExists = true;
+	        				userGroupRepository.delete(existingUserGroupRelationship);
+	        				break;
+	        			}
+	        			
+	        			
+	        		}
+	        		
+	        		if(!adminExists && groupMembers.size() > 1)
+	        		{
+		        		for(GroupMember groupMember:groupMembers)
+		        		{
+		        			if(groupMember.getId().equalsIgnoreCase(userId))
+		        			{
+		        				continue;
+		        			}
+		        			else
+		        			{
+	        				groupMember.setRole(Constants.GROUP_ADMIN_ROLE);
+	        				addUser(groupId,groupMember.getId(),Constants.GROUP_ADMIN_ROLE,userId);
+	        				userGroupRepository.delete(existingUserGroupRelationship);
+	        				break;
+		        			}
+	        			}
+	        		}
+	        		
+	        		if(groupMembers.size()==1)
+	        		{
+	        			deleteGroup(groupId);
+	        		}
+	        	}
+	        }
+	}
 }
